Refactor Auth code to package
