+++ dialogflow-fun-api/src/main/java/fr/lengrand/dialogflowfunapi/openbankproject/auth/Auth.java
@@ -1,7 +1,6 @@
-package fr.lengrand.dialogflowfunapi.openbankproject;
+package fr.lengrand.dialogflowfunapi.openbankproject.auth;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
